Edited version of the main.py file
The edits on this file take the external calls to the os and mpg123 to play the audio responses on the Alexa Pi. I found that the playback was getting cut of anytime the local wifi network was heavily loaded. I have switched to a python library called pygame and now there is no calling of external programs to play back the Alexa Responses.
This does add an additional dependency that you need to install a new python library using the following command
sudo apt-get install python-pygame